requirements.txt
streamlit==1.36.0
langchain_google_vertexai== 2.0.9
google-cloud-core==2.4.1
google-cloud-speech==2.28.1
google-cloud-storage==2.19.0
google-cloud-aiplatform==1.71.1
python-dotenv==1.0.1
langchain==0.3.3
langchain-community==0.3.2
langchain-core>=0.3.15
cryptography==44.0.0
vertexai==1.71.1

